+++ polkadot/runtime/common/src/assigned_slots/mod.rs
@@ -23,6 +23,7 @@
 //! This pallet should not be used on a production relay chain,
 //! only on a test relay chain (e.g. Rococo).
 
+#![cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 pub mod migration;
 
@@ -626,837 +627,8 @@ impl<T: Config> Pallet<T> {
 	}
 }
 
-/// tests for this pallet
 #[cfg(test)]
-mod tests {
-	use super::*;
+mod mock;
 
-	use crate::{assigned_slots, mock::TestRegistrar, slots};
-	use frame_support::{assert_noop, assert_ok, derive_impl, parameter_types};
-	use frame_system::EnsureRoot;
-	use pallet_balances;
-	use polkadot_primitives::BlockNumber;
-	use polkadot_primitives_test_helpers::{dummy_head_data, dummy_validation_code};
-	use polkadot_runtime_parachains::{
-		configuration as parachains_configuration, paras as parachains_paras,
-		shared as parachains_shared,
-	};
-	use sp_core::H256;
-	use sp_runtime::{
-		traits::{BlakeTwo256, IdentityLookup},
-		transaction_validity::TransactionPriority,
-		BuildStorage,
-		DispatchError::BadOrigin,
-	};
-
-	type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-	type Block = frame_system::mocking::MockBlockU32<Test>;
-
-	frame_support::construct_runtime!(
-		pub enum Test
-		{
-			System: frame_system,
-			Balances: pallet_balances,
-			Configuration: parachains_configuration,
-			ParasShared: parachains_shared,
-			Parachains: parachains_paras,
-			Slots: slots,
-			AssignedSlots: assigned_slots,
-		}
-	);
-
-	impl<C> frame_system::offchain::CreateTransactionBase<C> for Test
-	where
-		RuntimeCall: From<C>,
-	{
-		type Extrinsic = UncheckedExtrinsic;
-		type RuntimeCall = RuntimeCall;
-	}
-
-	impl<C> frame_system::offchain::CreateInherent<C> for Test
-	where
-		RuntimeCall: From<C>,
-	{
-		fn create_inherent(call: Self::RuntimeCall) -> Self::Extrinsic {
-			UncheckedExtrinsic::new_bare(call)
-		}
-	}
-
-	#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-	impl frame_system::Config for Test {
-		type BaseCallFilter = frame_support::traits::Everything;
-		type BlockWeights = ();
-		type BlockLength = ();
-		type RuntimeOrigin = RuntimeOrigin;
-		type RuntimeCall = RuntimeCall;
-		type Nonce = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Block = Block;
-		type RuntimeEvent = RuntimeEvent;
-		type DbWeight = ();
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = pallet_balances::AccountData<u64>;
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = ();
-		type OnSetCode = ();
-		type MaxConsumers = frame_support::traits::ConstU32<16>;
-	}
-
-	#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-	impl pallet_balances::Config for Test {
-		type AccountStore = System;
-	}
-
-	impl parachains_configuration::Config for Test {
-		type WeightInfo = parachains_configuration::TestWeightInfo;
-	}
-
-	parameter_types! {
-		pub const ParasUnsignedPriority: TransactionPriority = TransactionPriority::max_value();
-	}
-
-	impl parachains_paras::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type WeightInfo = parachains_paras::TestWeightInfo;
-		type UnsignedPriority = ParasUnsignedPriority;
-		type QueueFootprinter = ();
-		type NextSessionRotation = crate::mock::TestNextSessionRotation;
-		type OnNewHead = ();
-		type AssignCoretime = ();
-	}
-
-	impl parachains_shared::Config for Test {
-		type DisabledValidators = ();
-	}
-
-	parameter_types! {
-		pub const LeasePeriod: BlockNumber = 3;
-		pub static LeaseOffset: BlockNumber = 0;
-		pub const ParaDeposit: u64 = 1;
-	}
-
-	impl slots::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type Currency = Balances;
-		type Registrar = TestRegistrar<Test>;
-		type LeasePeriod = LeasePeriod;
-		type LeaseOffset = LeaseOffset;
-		type ForceOrigin = EnsureRoot<Self::AccountId>;
-		type WeightInfo = crate::slots::TestWeightInfo;
-	}
-
-	parameter_types! {
-		pub const PermanentSlotLeasePeriodLength: u32 = 3;
-		pub const TemporarySlotLeasePeriodLength: u32 = 2;
-		pub const MaxTemporarySlotPerLeasePeriod: u32 = 2;
-	}
-
-	impl assigned_slots::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type AssignSlotOrigin = EnsureRoot<Self::AccountId>;
-		type Leaser = Slots;
-		type PermanentSlotLeasePeriodLength = PermanentSlotLeasePeriodLength;
-		type TemporarySlotLeasePeriodLength = TemporarySlotLeasePeriodLength;
-		type MaxTemporarySlotPerLeasePeriod = MaxTemporarySlotPerLeasePeriod;
-		type WeightInfo = crate::assigned_slots::TestWeightInfo;
-	}
-
-	// This function basically just builds a genesis storage key/value store according to
-	// our desired mock up.
-	pub fn new_test_ext() -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::<Test>::default().build_storage().unwrap();
-		pallet_balances::GenesisConfig::<Test> {
-			balances: vec![(1, 10), (2, 20), (3, 30), (4, 40), (5, 50), (6, 60)],
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		crate::assigned_slots::GenesisConfig::<Test> {
-			max_temporary_slots: 6,
-			max_permanent_slots: 2,
-			_config: Default::default(),
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		t.into()
-	}
-
-	fn run_to_block(n: BlockNumber) {
-		while System::block_number() < n {
-			let mut block = System::block_number();
-			// on_finalize hooks
-			AssignedSlots::on_finalize(block);
-			Slots::on_finalize(block);
-			Parachains::on_finalize(block);
-			ParasShared::on_finalize(block);
-			Configuration::on_finalize(block);
-			Balances::on_finalize(block);
-			System::on_finalize(block);
-			// Set next block
-			System::set_block_number(block + 1);
-			block = System::block_number();
-			// on_initialize hooks
-			System::on_initialize(block);
-			Balances::on_initialize(block);
-			Configuration::on_initialize(block);
-			ParasShared::on_initialize(block);
-			Parachains::on_initialize(block);
-			Slots::on_initialize(block);
-			AssignedSlots::on_initialize(block);
-		}
-	}
-
-	#[test]
-	fn basic_setup_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-			assert_eq!(AssignedSlots::current_lease_period_index(), 0);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-
-			run_to_block(3);
-			assert_eq!(AssignedSlots::current_lease_period_index(), 1);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_fails_for_unknown_para() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-				),
-				Error::<Test>::ParaDoesntExist
-			);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_fails_for_invalid_origin() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::signed(1),
-					ParaId::from(1_u32),
-				),
-				BadOrigin
-			);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_fails_when_not_parathread() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-			assert_ok!(TestRegistrar::<Test>::make_parachain(ParaId::from(1_u32)));
-
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-				),
-				Error::<Test>::NotParathread
-			);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_fails_when_existing_lease() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			// Register lease in current lease period
-			assert_ok!(Slots::lease_out(ParaId::from(1_u32), &1, 1, 1, 1));
-			// Try to assign a perm slot in current period fails
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-				),
-				Error::<Test>::OngoingLeaseExists
-			);
-
-			// Cleanup
-			assert_ok!(Slots::clear_all_leases(RuntimeOrigin::root(), 1.into()));
-
-			// Register lease for next lease period
-			assert_ok!(Slots::lease_out(ParaId::from(1_u32), &1, 1, 2, 1));
-			// Should be detected and also fail
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-				),
-				Error::<Test>::OngoingLeaseExists
-			);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_fails_when_max_perm_slots_exceeded() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				2,
-				ParaId::from(2_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				3,
-				ParaId::from(3_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_ok!(AssignedSlots::assign_perm_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-			));
-			assert_ok!(AssignedSlots::assign_perm_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(2_u32),
-			));
-			assert_eq!(assigned_slots::PermanentSlotCount::<Test>::get(), 2);
-
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(3_u32),
-				),
-				Error::<Test>::MaxPermanentSlotsExceeded
-			);
-		});
-	}
-
-	#[test]
-	fn assign_perm_slot_succeeds_for_parathread() {
-		new_test_ext().execute_with(|| {
-			let mut block = 1;
-			run_to_block(block);
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_eq!(assigned_slots::PermanentSlotCount::<Test>::get(), 0);
-			assert_eq!(assigned_slots::PermanentSlots::<Test>::get(ParaId::from(1_u32)), None);
-
-			assert_ok!(AssignedSlots::assign_perm_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-			));
-
-			// Para is a lease holding parachain for PermanentSlotLeasePeriodLength * LeasePeriod
-			// blocks
-			while block < 9 {
-				println!("block #{}", block);
-
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-
-				assert_eq!(assigned_slots::PermanentSlotCount::<Test>::get(), 1);
-				assert_eq!(AssignedSlots::has_permanent_slot(ParaId::from(1_u32)), true);
-				assert_eq!(
-					assigned_slots::PermanentSlots::<Test>::get(ParaId::from(1_u32)),
-					Some((0, 3))
-				);
-
-				assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 2), true);
-
-				block += 1;
-				run_to_block(block);
-			}
-
-			// Para lease ended, downgraded back to parathread (on-demand parachain)
-			assert_eq!(TestRegistrar::<Test>::is_parathread(ParaId::from(1_u32)), true);
-			assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 5), false);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_fails_for_unknown_para() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-					SlotLeasePeriodStart::Current
-				),
-				Error::<Test>::ParaDoesntExist
-			);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_fails_for_invalid_origin() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::signed(1),
-					ParaId::from(1_u32),
-					SlotLeasePeriodStart::Current
-				),
-				BadOrigin
-			);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_fails_when_not_parathread() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-			assert_ok!(TestRegistrar::<Test>::make_parachain(ParaId::from(1_u32)));
-
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-					SlotLeasePeriodStart::Current
-				),
-				Error::<Test>::NotParathread
-			);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_fails_when_existing_lease() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			// Register lease in current lease period
-			assert_ok!(Slots::lease_out(ParaId::from(1_u32), &1, 1, 1, 1));
-			// Try to assign a perm slot in current period fails
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-					SlotLeasePeriodStart::Current
-				),
-				Error::<Test>::OngoingLeaseExists
-			);
-
-			// Cleanup
-			assert_ok!(Slots::clear_all_leases(RuntimeOrigin::root(), 1.into()));
-
-			// Register lease for next lease period
-			assert_ok!(Slots::lease_out(ParaId::from(1_u32), &1, 1, 2, 1));
-			// Should be detected and also fail
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(1_u32),
-					SlotLeasePeriodStart::Current
-				),
-				Error::<Test>::OngoingLeaseExists
-			);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_fails_when_max_temp_slots_exceeded() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			// Register 6 paras & a temp slot for each
-			for n in 0..=5 {
-				assert_ok!(TestRegistrar::<Test>::register(
-					n,
-					ParaId::from(n as u32),
-					dummy_head_data(),
-					dummy_validation_code()
-				));
-
-				assert_ok!(AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(n as u32),
-					SlotLeasePeriodStart::Current
-				));
-			}
-
-			assert_eq!(assigned_slots::TemporarySlotCount::<Test>::get(), 6);
-
-			// Attempt to assign one more temp slot
-			assert_ok!(TestRegistrar::<Test>::register(
-				7,
-				ParaId::from(7_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-			assert_noop!(
-				AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(7_u32),
-					SlotLeasePeriodStart::Current
-				),
-				Error::<Test>::MaxTemporarySlotsExceeded
-			);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_succeeds_for_single_parathread() {
-		new_test_ext().execute_with(|| {
-			let mut block = 1;
-			run_to_block(block);
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_eq!(assigned_slots::TemporarySlots::<Test>::get(ParaId::from(1_u32)), None);
-
-			assert_ok!(AssignedSlots::assign_temp_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-				SlotLeasePeriodStart::Current
-			));
-			assert_eq!(assigned_slots::TemporarySlotCount::<Test>::get(), 1);
-			assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 1);
-
-			// Block 1-5
-			// Para is a lease holding parachain for TemporarySlotLeasePeriodLength * LeasePeriod
-			// blocks
-			while block < 6 {
-				println!("block #{}", block);
-				println!("lease period #{}", AssignedSlots::current_lease_period_index());
-				println!("lease {:?}", slots::Leases::<Test>::get(ParaId::from(1_u32)));
-
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-
-				assert_eq!(AssignedSlots::has_temporary_slot(ParaId::from(1_u32)), true);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 1);
-				assert_eq!(
-					assigned_slots::TemporarySlots::<Test>::get(ParaId::from(1_u32)),
-					Some(ParachainTemporarySlot {
-						manager: 1,
-						period_begin: 0,
-						period_count: 2, // TemporarySlotLeasePeriodLength
-						last_lease: Some(0),
-						lease_count: 1
-					})
-				);
-
-				assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 1), true);
-
-				block += 1;
-				run_to_block(block);
-			}
-
-			// Block 6
-			println!("block #{}", block);
-			println!("lease period #{}", AssignedSlots::current_lease_period_index());
-			println!("lease {:?}", slots::Leases::<Test>::get(ParaId::from(1_u32)));
-
-			// Para lease ended, downgraded back to on-demand parachain
-			assert_eq!(TestRegistrar::<Test>::is_parathread(ParaId::from(1_u32)), true);
-			assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 3), false);
-			assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 0);
-
-			// Block 12
-			// Para should get a turn after TemporarySlotLeasePeriodLength * LeasePeriod blocks
-			run_to_block(12);
-			println!("block #{}", block);
-			println!("lease period #{}", AssignedSlots::current_lease_period_index());
-			println!("lease {:?}", slots::Leases::<Test>::get(ParaId::from(1_u32)));
-
-			assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-			assert_eq!(Slots::already_leased(ParaId::from(1_u32), 4, 5), true);
-			assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 1);
-		});
-	}
-
-	#[test]
-	fn assign_temp_slot_succeeds_for_multiple_parathreads() {
-		new_test_ext().execute_with(|| {
-			// Block 1, Period 0
-			run_to_block(1);
-
-			// Register 6 paras & a temp slot for each
-			// (3 slots in current lease period, 3 in the next one)
-			for n in 0..=5 {
-				assert_ok!(TestRegistrar::<Test>::register(
-					n,
-					ParaId::from(n as u32),
-					dummy_head_data(),
-					dummy_validation_code()
-				));
-
-				assert_ok!(AssignedSlots::assign_temp_parachain_slot(
-					RuntimeOrigin::root(),
-					ParaId::from(n as u32),
-					if (n % 2).is_zero() {
-						SlotLeasePeriodStart::Current
-					} else {
-						SlotLeasePeriodStart::Next
-					}
-				));
-			}
-
-			// Block 1-5, Period 0-1
-			for n in 1..=5 {
-				if n > 1 {
-					run_to_block(n);
-				}
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), false);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-
-			// Block 6-11, Period 2-3
-			for n in 6..=11 {
-				run_to_block(n);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), false);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-
-			// Block 12-17, Period 4-5
-			for n in 12..=17 {
-				run_to_block(n);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), true);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-
-			// Block 18-23, Period 6-7
-			for n in 18..=23 {
-				run_to_block(n);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), false);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-
-			// Block 24-29, Period 8-9
-			for n in 24..=29 {
-				run_to_block(n);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), false);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-
-			// Block 30-35, Period 10-11
-			for n in 30..=35 {
-				run_to_block(n);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(0)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(2_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(3_u32)), false);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(4_u32)), true);
-				assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(5_u32)), true);
-				assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 2);
-			}
-		});
-	}
-
-	#[test]
-	fn unassign_slot_fails_for_unknown_para() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::unassign_parachain_slot(RuntimeOrigin::root(), ParaId::from(1_u32),),
-				Error::<Test>::SlotNotAssigned
-			);
-		});
-	}
-
-	#[test]
-	fn unassign_slot_fails_for_invalid_origin() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::assign_perm_parachain_slot(
-					RuntimeOrigin::signed(1),
-					ParaId::from(1_u32),
-				),
-				BadOrigin
-			);
-		});
-	}
-
-	#[test]
-	fn unassign_perm_slot_succeeds() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_ok!(AssignedSlots::assign_perm_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-			));
-
-			assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-
-			assert_ok!(AssignedSlots::unassign_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-			));
-
-			assert_eq!(assigned_slots::PermanentSlotCount::<Test>::get(), 0);
-			assert_eq!(AssignedSlots::has_permanent_slot(ParaId::from(1_u32)), false);
-			assert_eq!(assigned_slots::PermanentSlots::<Test>::get(ParaId::from(1_u32)), None);
-
-			assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 2), false);
-		});
-	}
-
-	#[test]
-	fn unassign_temp_slot_succeeds() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code(),
-			));
-
-			assert_ok!(AssignedSlots::assign_temp_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-				SlotLeasePeriodStart::Current
-			));
-
-			assert_eq!(TestRegistrar::<Test>::is_parachain(ParaId::from(1_u32)), true);
-
-			assert_ok!(AssignedSlots::unassign_parachain_slot(
-				RuntimeOrigin::root(),
-				ParaId::from(1_u32),
-			));
-
-			assert_eq!(assigned_slots::TemporarySlotCount::<Test>::get(), 0);
-			assert_eq!(assigned_slots::ActiveTemporarySlotCount::<Test>::get(), 0);
-			assert_eq!(AssignedSlots::has_temporary_slot(ParaId::from(1_u32)), false);
-			assert_eq!(assigned_slots::TemporarySlots::<Test>::get(ParaId::from(1_u32)), None);
-
-			assert_eq!(Slots::already_leased(ParaId::from(1_u32), 0, 1), false);
-		});
-	}
-	#[test]
-	fn set_max_permanent_slots_fails_for_no_root_origin() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::set_max_permanent_slots(RuntimeOrigin::signed(1), 5),
-				BadOrigin
-			);
-		});
-	}
-	#[test]
-	fn set_max_permanent_slots_succeeds() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_eq!(MaxPermanentSlots::<Test>::get(), 2);
-			assert_ok!(AssignedSlots::set_max_permanent_slots(RuntimeOrigin::root(), 10),);
-			assert_eq!(MaxPermanentSlots::<Test>::get(), 10);
-		});
-	}
-
-	#[test]
-	fn set_max_temporary_slots_fails_for_no_root_origin() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_noop!(
-				AssignedSlots::set_max_temporary_slots(RuntimeOrigin::signed(1), 5),
-				BadOrigin
-			);
-		});
-	}
-	#[test]
-	fn set_max_temporary_slots_succeeds() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_eq!(MaxTemporarySlots::<Test>::get(), 6);
-			assert_ok!(AssignedSlots::set_max_temporary_slots(RuntimeOrigin::root(), 12),);
-			assert_eq!(MaxTemporarySlots::<Test>::get(), 12);
-		});
-	}
-}
+#[cfg(test)]
+mod tests;
